requirements.txt
openai>=1.40.0
numpy>=1.26
python-dotenv>=1.0.1
tiktoken>=0.7.0
